@@ -38,9 +38,9 @@ func (t *SlicerToxic) chunk(start int, end int) []int {
 	}
 
 	mid := start + (end-start)/2
-	//#nosec
+
 	if t.SizeVariation > 0 {
-		mid += rand.Intn(t.SizeVariation*2) - t.SizeVariation
+		mid += rand.Intn(t.SizeVariation*2) - t.SizeVariation // #nosec G404 -- was ignored before too
 	}
 	left := t.chunk(start, mid)
 	right := t.chunk(mid, end)
